fix: disable event viewer cache
